Split DNSPoller into DNSPoller and RuleGen
The rule management and regeneration is now in RuleGen, while DNSPoller
is only the simpler DNS poll component. DNSPoller still polls all FQDN
names managed by RuleGen and relies on it directly.
